Finished up name prompts to some what work at least
Added check name which we can later fill up by restrictions
Added basic requests to the server after the name is checked.
Added ReusedButtonCtrl so that every controller that shares the back and sound buttons don't have to have functions.
I'm unsure about the sendGet function as there is probably a prettier way to do it with rest. For now it kinda works